remove line search

+++ solvers/MLP.py
@@ -190,7 +190,7 @@ class MLP:
 
         # Vectorized evaluation of g
         distrubed_output_terminal_flat = g(disturbed_input_terminal_flat)  # Evaluate disturbed terminal condition, shape (batch_size * MC_g, 1)
-        self.evaluation_counter += MC_g
+        self.evaluation_counter+=MC_g
 
         # Reshape back to (batch_size, MC_g, 1)
         distrubed_output_terminal =  distrubed_output_terminal_flat.reshape(batch_size, MC_g, 1)
@@ -242,7 +242,7 @@ class MLP:
                 simulated_u_flat = simulated_u.reshape(-1,1)
                 simulated_z_flat = simulated_z.reshape(-1, dim)
                 y_flat = f(input_intermediates_flat, simulated_u_flat, simulated_z_flat)
-                self.evaluation_counter += MC_f
+                self.evaluation_counter+=MC_f
                 y = y_flat.reshape(batch_size, MC_f, 1)
 
                 u += wloc[:, k,q-1][:, jnp.newaxis] * jnp.mean(y, axis=1)
@@ -263,7 +263,7 @@ class MLP:
                     simulated_u_flat = simulated_u.reshape(-1, 1)
                     simulated_z_flat = simulated_z.reshape(-1, dim)
                     y_flat = f(input_intermediates_flat, simulated_u_flat, simulated_z_flat)
-                    self.evaluation_counter += MC_f
+                    self.evaluation_counter+=MC_f
                     y = y_flat.reshape(batch_size, MC_f, 1)
 
                     u -= wloc[:, k, q - 1][:, jnp.newaxis] * jnp.mean(y, axis=1)  # Adjust u values
@@ -271,7 +271,7 @@ class MLP:
                     z -= wloc[:, k, q - 1][:, jnp.newaxis] * jnp.sum(y * W, axis=1) / (MC_f * delta_t)  # Adjust z values
         output_cated = jnp.concatenate((u, z), axis=-1)  # Concatenate adjusted u and z values, shape (batch_size, dim + 1)
         norm_estimation = self.equation.norm_estimation
-        return jnp.clip(output_cated, -norm_estimation, norm_estimation)  # Clip the output to avoid numerical instability
+        return jnp.clip(output_cated, -norm_estimation, norm_estimation).astype(jnp.float16)  # Clip the output to avoid numerical instability
 
     def u_solve(self, n, rho, x_t):
         '''
